requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aiosignal==1.3.1
alchemy==20.5
alembic==1.13.2
amqp==5.2.0
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
Babel==2.8.0
bcrypt==4.1.3
billiard==4.2.0
blinker==1.8.2
celery==5.4.0
certifi==2024.7.4
chardet==3.0.4
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
docutils==0.21.2
email_validator==2.2.0
Faker==26.0.0
filelock==3.0.12
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.4.1
greenlet==3.0.3
httplib2==0.20.2
idna==2.8
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
kombu==5.3.7
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==2.1.5
more-itertools==8.10.0
multidict==6.0.5
netifaces==0.11.0
oauthlib==3.2.0
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
python-daemon==2.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2022.1
PyYAML==5.4.1
redis==5.0.7
requests==2.22.0
SecretStorage==3.3.1
simplejson==3.17.6
six==1.16.0
SQLAlchemy==2.0.31
systemd-python==234
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.25.11
vine==5.1.0
wadllib==1.3.6
wcwidth==0.2.13
Werkzeug==3.0.3
WTForms==3.1.2
yarl==1.9.4
zipp==1.0.0
